requirements.txt
amqp==5.1.1
aniso8601==9.0.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Faker==19.4.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
Werkzeug==2.3.7
google-cloud-storage==2.13.0
google-cloud-pubsub==2.18.4